refactor: refactor folder structure and files name
