+++ Components/page-sections/layout.js
@@ -2,11 +2,9 @@ import Navbar from "./navbar";
 
 export default function Layout({ children }) {
   return (
-
     <div className="flex flex-col min-h-screen">
       <Navbar />
       {children}
-
     </div>
   );
 }
