+++ Components/search-result/search-result.js
