+++ Components/view-cinema-page/hero-section.js
