+++ Components/view-cinema-page/movies-section.js
