+++ Components/view-movie-page/cinemas-section.js
