+++ pages/_app.js
@@ -1,4 +1,4 @@
-import Layout from "@/Components/layout";
+import Layout from "@/Components/page-sections/layout";
 import "@/styles/globals.css";
 import { Provider } from "@/components/ui/provider";
 
@@ -6,7 +6,7 @@ export default function App({ Component, pageProps }) {
   return (
     <Provider>
       <Layout>
-      <Component {...pageProps} />
+        <Component {...pageProps} />
       </Layout>
     </Provider>
   );
