+++ pages/api/auth/forgot-password.js
@@ -1,5 +1,5 @@
 import { supabase } from "@/lib/supabase-client"; // เชื่อมต่อกับ Supabase
-import db from "../../utils/db"; // เชื่อมต่อกับฐานข้อมูล PostgreSQL
+import db from "../../../utils/db"; // เชื่อมต่อกับฐานข้อมูล PostgreSQL
 
 export default async function handler(req, res) {
   if (req.method !== "POST") {
