+++ pages/api/auth/logout.js
