+++ pages/api/auth/register.js
