+++ pages/index.js
@@ -1,11 +1,11 @@
 import { useState } from "react";
-import Filter from "@/Components/filter.js";
-import MoviesCard from "@/Components/movieCard.js";
-import MoviesAffiliate from "@/Components/moviesAffilate.js";
-import ResultFromFilter from "@/Components/resultFromFilter.js";
+import Filter from "@/Components/landing-page/search-filter-movie.js";
+import MoviesCard from "@/Components/landing-page/movie-card.js";
+import MoviesAffiliate from "@/Components/landing-page/movies-affilate.js";
+import ResultFromFilter from "@/Components/search-result/search-result.js";
 import Head from "next/head";
-import Footer from "@/Components/footer.js";
-import Navbar from "@/Components/navbar";
+import Footer from "@/Components/page-sections/footer.js";
+import Navbar from "@/Components/page-sections/navbar";
 
 export default function Home() {
   const [filteredMovies, setFilteredMovies] = useState([]);
